Redirect to canonical URLs
URLs referring to a directory should always end in a slash. URLs
referring to a file should never end in a slash. We now redirect to the
correct form.

The reason this is important is that incorrect URLs can result to broken
relative links. For example, if the page at gemini://example.org/gemlog/
has a relative link to post.gmi, and we drop the trailing slash (so it
ends with "gemlog", instead of "gemlog/"), the link to post.gmi would
refer to "gemini://example.org/post.gmi" instead of
"gemini://example.org/gemlog/post.gmi".